feat: Update ESLint settings and add new modal functionality
- Updated ESLint settings to include additional file types and validate specific languages.
- Added new modal functionality to display modals with customizable title, body, footer, and buttons.
- The modal can be closed by clicking on the overlay or pressing the escape key.
- The modal component can be created and destroyed programmatically using the `$modal.create` and `$modal.destroy` methods.
- Added a confirm modal method that prompts the user with a title and message and returns a promise that resolves to a boolean value.

+++ frontend/src/components/InvitationList/InvitationItem.vue
@@ -50,7 +50,9 @@ export default defineComponent({
             this.$toast.info(this.__("Copied to clipboard"));
         },
         async deleteLocalInvitation() {
-            await this.deleteInvitation(this.invite.id);
+            if (await this.$modal.confirm(this.__("Are you sure?"), this.__("Do you really want to delete this invitation?"))) {
+                await this.deleteInvitation(this.invite.id);
+            }
         },
         ...mapActions(useInvitationStore, ["deleteInvitation"]),
     },
